# docs/contracts/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOUL V1 API",
    "version": "1.0.0",
    "description": "Machine-readable companion to docs/FLUTTER_API_HANDOFF.md."
  },
  "servers": [
    {
      "url": "http://localhost/api/v1",
      "description": "Local development; replace the origin per environment"
    }
  ],
  "paths": {
...
    "/catalogs/profile": {
      "get": {
        "operationId": "api.v1.catalogs.profile",
        "summary": "Localized interests, traits and help categories",
        "tags": [
          "Catalogs"
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
    "/verification/summary": {
      "get": {
        "operationId": "api.v1.verification.summary",
        "summary": "Separate account checks and public badge states",
        "tags": [
          "Verification"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "2XX": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum token"
      },
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session"
      },
      "cloudinarySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Cld-Signature"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data",
          "meta"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "Success": {
        "description": "Successful SOUL response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SuccessEnvelope"
            }
          }
        }
      },
      "Error": {
        "description": "Standard SOUL error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}